Move all Trim methods to dedicated file
